Add timestamp measurement to form submission to protect against spambots.
Needed some PHP in the form page, so rooms.html now redirects to apply.php,
which contains the actual form data. Timestamp in the form is hashed using
the same library used to create unique URLs for applications. If the difference
between form display and form submission is less than 15 seconds, form will
be rejected.